+++ app/src/main/java/open/furaffinity/client/fragmentTabsNew/manageBadges.java
@@ -1,0 +1,33 @@
+package open.furaffinity.client.fragmentTabsNew;
+
+import android.view.View;
+
+import open.furaffinity.client.R;
+
+public class manageBadges extends open.furaffinity.client.abstractClasses.tabFragment {
+
+    @Override
+    protected int getLayout() {
+        return R.layout.not_implemented_yet;
+    }
+
+    @Override
+    protected void getElements(View rootView) {
+
+    }
+
+    @Override
+    protected void initPages() {
+
+    }
+
+    @Override
+    protected void fetchPageData() {
+
+    }
+
+    @Override
+    protected void updateUIElementListeners(View rootView) {
+
+    }
+}
