--- app/src/main/java/open/furaffinity/client/fragmentTabsOld/manageBadges.java
@@ -1,22 +1,0 @@
-package open.furaffinity.client.fragmentTabsOld;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.fragment.app.Fragment;
-
-import open.furaffinity.client.R;
-
-public class manageBadges extends Fragment {
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.not_implemented_yet, container, false);
-        return rootView;
-    }
-}
